pass invalid_operation.slt

+++ src/frontend/src/handler/drop_index.rs
@@ -55,9 +55,19 @@ pub async fn handle_drop_index(
                 };
                 return match reader.get_table_by_name(db_name, schema_path, &index_name) {
                     Ok((table, _)) => match table.table_type() {
-                        TableType::Table => Err(RwError::from(ErrorCode::InvalidInputSyntax(
-                            "Use `DROP TABLE` to drop a table.".to_owned(),
-                        ))),
+                        TableType::Table => {
+                            // TODO(Yuanxin): Remove this after unsupporting `CREATE MATERIALIZED
+                            // SOURCE`.
+                            if table.associated_source_id().is_some() {
+                                Err(RwError::from(ErrorCode::InvalidInputSyntax(
+                                    "Use `DROP SOURCE` to drop a source.".to_owned(),
+                                )))
+                            } else {
+                                Err(RwError::from(ErrorCode::InvalidInputSyntax(
+                                    "Use `DROP TABLE` to drop a table.".to_owned(),
+                                )))
+                            }
+                        }
                         TableType::MaterializedView => {
                             Err(RwError::from(ErrorCode::InvalidInputSyntax(
                                 "Use `DROP MATERIALIZED VIEW` to drop a materialized view."
